requirements.txt
fastapi==0.118.2
uvicorn==0.37.0
python-dotenv==1.1.1
langchain==0.3.27
langchain-openai==0.3.35
langchain-chroma==0.2.6
langchain-community==0.3.31
langchain-core==0.3.78
langchain-text-splitters==0.3.11
openai==2.2.0
chromadb==1.1.1
twilio==9.8.3
aiohttp==3.13.0
python-multipart==0.0.20
